rmeza-SHP-07 included stories SHP-08 and SHP-09, also added some dummy response for visualizing functionability

+++ BackEnd/src/main/java/com/javaschool/modelmapper/TransportType.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class TransportType {
     private int id;
     private String description;
-    private int pricePerMile;
+    private double pricePerMile;
 }
